+++ kit/ag-grid/index.js
@@ -13,8 +13,6 @@ import compareVersions from 'compare-versions';
  * This allows applications to provide Hoist with their duly imported and licensed
  * versions of ag-Grid.  In particular, note that many (but not all) users of Hoist
  * will be expected to have the enterprise version of ag-Grid.
- *
- * See the file agGrid.js in the application's root directory.
  */
 export let AgGridReact = null;
 export let AgGridUtils = null;
@@ -25,8 +23,9 @@ const MAX_VERSION = '23.*.*';
 
 /**
  * Expose application versions of ag-Grid to Hoist.
+ * Typically called in the Bootstrap.js. of the application.
  */
-export function installAgGridImpls(ComponentReactWrapper, Utils, version) {
+export function installAgGrid(ComponentReactWrapper, Utils, version) {
     if (compareVersions(version, MIN_VERSION) < 0 || compareVersions(version, MAX_VERSION) > 0) {
         console.error(
             `This version of Hoist requires an ag-Grid version between ${MIN_VERSION} and ` +