+++ CHANGELOG.md
@@ -46,9 +46,9 @@ wide variety of enterprise software projects. For any questions regarding this c
    full-screen component.
 * `PinPad` and `PinPadModel` have been moved to `@xh/hoist/cmp/pinpad`, and is now available for use
   with both standard and mobile toolkits.
-* Grid dependencies updated to properly reflect application-level licensing requirements.
-  Applications must now import and provide their licensed version of ag-Grid to Hoist.  See
-  file `Bootstrap.js` in Toolbox for an example.
+* Third-party dependencies updated to properly reflect application-level licensing requirements.
+  Applications must now import and provide their licensed version of ag-Grid, and Highcharts to
+  Hoist.  See file `Bootstrap.js` in Toolbox for an example.
 
 [Commit Log](https://github.com/xh/hoist-react/compare/v34.0.0...develop)